requirements.txt
transformers
langchain
langchain-huggingface
torch
pandas
python-dotenv

